requirements.txt
pymupdf>=1.24.0
python-pptx>=0.6.23
Pillow>=10.0.0
flet>=0.22.0
PyInstaller>=6.15.0
